+++ docs/changelog.md
@@ -1,5 +1,9 @@
 ## Changelog
 
+## Version 3.0.11
+- *Sequencer* - Made PIXI fix settings default to being turned off
+- *Effects* - Fixed some persistent effects not being deleted from scenes after their attached sources were deleted
+
 ## Version 3.0.10
 - *Sequencer* - Fixed Effect Player not working in Foundry v11
 - *Sequencer* - Fixed private modules showing up in the Database Viewer